Buildings etc
Added Buildings
MathUtils changes
GetResourceString added to resource database
Resource Identifier now allows float ammounts

+++ Assets/Scripts/Data Store/ResourceIndentifier.cs
@@ -8,5 +8,5 @@ using System.Threading.Tasks;
 public struct ResourceIndentifier
 {
 	public int id;
-	public int ammount;
+	public float ammount;
 }
